Finish basic back-end
At this point the game allows user to play tic-tac-toe, and a winner is decided after 3 moves.This step was made to check the logic behind the game and the syntax of ruby.

Obs: The game still requires work in the back-end, so as to allow players to play multiple times, display draw if there are not more available moves. Unallow the options of picking a square which was previously selected by the other user, etc.

Next step: basic front-end.